feat(SS): M131b money helper + parseDecimal (F68 locale-bug)
Centraliserar M131-formeln i src/shared/money.ts och ersätter grep-
baserad enforcement med AST-scan (scripts/check-m131-ast.mjs). Fixar
samtidigt produktionsbugg F68 där parseFloat("99,50") tolkades som
99 — svenska användare tappade decimalerna tyst i alla belopp-inputs.

money.ts:
- multiplyKrToOre(qty, priceKr) — kanonisk M131-formel
- multiplyDecimalByOre(qty, priceOre) — för main (pris redan i ore)
- parseDecimal(value) — locale-safe, konverterar "," → "." trim + NaN

F68 callsites migrerade (9 ställen / 7 filer):
- Invoice/ExpenseLineRow: qty + unit_price_kr
- PayExpenseDialog: fee + amount
- CustomerPriceTable, FixedAssetFormDialog (cost, residual, rate)
- DisposeDialog (ersätter inline .replace(",", "."))
- StepCompany/StepConfirm: share_capital

M131 AST-scan:
- check-m131-ast.mjs med self-test (7 cases) + produktionsscan
- Flaggar binär * på _kr-identifier utanför money.ts
- Fångar aliaser (const p = line.unit_price_kr; qty * p) som
  grep-versionen missade
- check:m131 pekar nu på AST-versionen; grep kvar som check:m131-grep
  backup

CLAUDE.md M131:
- Helper-referens + AST-enforcement dokumenterat
- Scope-not: F47-backlog avklarad (LineRow-komponenter migrerade)

Regressionstester:
- tests/shared/parse-decimal.test.ts — 8 cases inkl. dokumenterat
  gammalt parseFloat("99,50") === 99 beteende

+++ src/renderer/components/expenses/PayExpenseDialog.tsx
@@ -2,6 +2,7 @@ import { useState } from 'react'
 import type { ExpenseDetail } from '../../../shared/types'
 import { usePayExpense } from '../../lib/hooks'
 import { formatKr, toOre, toKr, todayLocal } from '../../lib/format'
+import { parseDecimal } from '../../../shared/money'
 
 interface PayExpenseDialogProps {
   expense: ExpenseDetail
@@ -31,10 +32,10 @@ export function PayExpenseDialog({
   const handleSubmit = async () => {
     setError(null)
     try {
-      const feeOre = bankFeeStr ? toOre(parseFloat(bankFeeStr)) : undefined
+      const feeOre = bankFeeStr ? toOre(parseDecimal(bankFeeStr)) : undefined
       await payMutation.mutateAsync({
         expense_id: expense.id,
-        amount_ore: toOre(parseFloat(amountKr)),
+        amount_ore: toOre(parseDecimal(amountKr)),
         payment_date: paymentDate,
         payment_method: paymentMethod,
         account_number: accountNumber,
